+++ Changes.md
@@ -51,6 +51,7 @@ Change Log in version 3.0.3 (2020073104)
 48. Fix 'Duplicate code in get_logo_title()' - #208.
 49. Fix 'responsivealerts used?' - #205.
 50. Fix 'fonttitlecolorcourse setting not used' - #154.
+51. Fix 'enablealertcoursepages setting used?' - #151.
 
 Change Log in version 3.0.2 (2020073103)
 ========================================